Refactor admin list tables into trait so that they are re-usable.
